+++ boards/INV/Inc/Io/Io_TimerRotorGen.h
@@ -1,4 +1,4 @@
 #pragma once
 
-void   Io_TimerRotorGen_SetFundFreq(double fund_frequency);
-double Io_TimerRotorGen_GetPosition(void);
+void  Io_TimerRotorGen_SetFundFreq(float fund_frequency);
+float Io_TimerRotorGen_GetPosition(void);
