+++ boards/INV/Src/App/controls/App_ParkClarkeTransform.c
@@ -10,9 +10,9 @@ struct PhaseValues parkClarkeTransform(
     float alpha = dqs_voltages->d * cosf(theta) - dqs_voltages->q * sinf(theta);
     float beta  = dqs_voltages->d * sinf(theta) + dqs_voltages->q * cosf(theta);
 
-    phase_voltages.a = (M_SQRT3 / 2) * alpha - beta / 2;
+    phase_voltages.a = ((float)M_SQRT3 / 2) * alpha - beta / 2;
     phase_voltages.b = beta;
-    phase_voltages.c = -(M_SQRT3 / 2) * alpha - beta / 2;
+    phase_voltages.c = -((float)M_SQRT3 / 2) * alpha - beta / 2;
 
     return phase_voltages;
 }
